Add precision fields to numerical question ui
Adds edit ui for "Answer with Precision" as a "Numerical Answer" subtype.

Precision answers are rounded to a precision of 16 to handle floating
point errors and then truncated to the specified precision for grading
and comparison.

If any of the potential answers to a numerical question is a precision
answer then the student form will accept precision answers. If none of
the possible answers are precision answers then the student form will
continue to round to 4 decimal places.

Refs CNVS-10866

Test plan:
 - Create a new question
 - Select a question type of "Numerical Answer"
 - Select answer type of "Answer with Precision"
 - You should see an answer input field and a precision input field
 - The precision field should default to 10 and accept input between
 1 and 16, these are the limits of javascript floats
 - The answer input should accept numerical inputs with or without an `e`
 exponent and be formatted according to the configured precision
 - The quiz submission form should accept answers with a precision of up
 to 16 in normal or exponential format (#.##e##)

Change-Id: Idc117a61040a43b038272afec9319bada43e5f54
Reviewed-on: https://gerrit.instructure.com/64068
Reviewed-by: Matt Berns <mberns@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Tested-by: Jenkins
Product-Review: Jason Sparks <jsparks@instructure.com>