Makes the moderation page load all students
This will make the page load all the students in batches of 50.
It will continuously load them all until they are all loaded.
Understandably this isn't the best for the DB, but it works well
in the context of accessibility (which lazy load on scroll) doesn't.

Once a load more button or the like is clarified for UX and/or we
see what impact this has in production we can make adjustments
to not auto-load if needed.

closes CNVS-23839

Test Plan:
  - Set up a moderated grading environment
  - Make sure you have > 50 students (200+ would be a great test)
  - You should go to the moderation page and see all the students
    listed.

Change-Id: Ia55b75bffc4f44325cb62064bc9cb6b7a239b655
Reviewed-on: https://gerrit.instructure.com/64612
Reviewed-by: Sterling Cobb <sterling@instructure.com>
Tested-by: Jenkins
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>