add button role to quiz edit page cancel link
Previously the 'Cancel' button on the quiz edit page
was being read as a link instead of a button. Added
a button role to the erb so it now reads as cancel
'button.' Attempting a new commit by request from
cwang.

fixes CNVS-23752

Test Plan:
1- Navigate to the Quiz Edit page with a screen reader.
2- Use the button hotkey to navigate through the page buttons.
The Cancel button should be read as a button with the proper description.

Change-Id: I3ef145915f5520f4de5604cbecaa041a28213a4f
Reviewed-on: https://gerrit.instructure.com/65039
Tested-by: Jenkins
Reviewed-by: Chris Wang <cwang@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Aaron Cannon <acannon@instructure.com>