Fix icon accessibility issue in To Do list
fixes CNVS-23619

It will probably be a good idea to make sure that the color that is
being used now is not an accessibility issue with Nathan. The
green check icon next to recent feedback items will always be green
now and is not meant to convey additional information.

Previously, To Do and Coming Up items had check marks next to them that
were gray and because of this a correlation existed among the three
lists that allowed color to convery additional information that
screenreader users did not have access to.

Test plan:
* Ensure the following for both a student and a teacher
* Ensure that in the To Do and Coming Up lists if the assignment is:
  a quiz: a quiz icon is shown
  a graded discussion: a graded discusion icon is shown
  an assignment: an assignment icon is shown
* Ensure that in the Coming Up list if an event is there a calendar icon
  is shown
* Ensure that in the Recent Feedback list the icon shown next to every
  item is a green checkmark
* Ensure that for each calendar event, quiz, discussion, and assignment
  icon a screen reader reads that the assignment is of that type

Change-Id: I12c435a2694921c1ade775a935105cf813e08c3d
Reviewed-on: https://gerrit.instructure.com/65690
Tested-by: Jenkins
QA-Review: Adrian Russell <arussell@instructure.com>
Reviewed-by: Matthew Wheeler <mwheeler@instructure.com>
Product-Review: Andrew Butterfield <abutterfield@instructure.com>