add lti post message for scrolling to top of iframe
fixes PLAT-1282 PLAT-1283

If a tool is a single page app and has resized the iframe
it will need to scroll to top when a new page is loaded

Sometimes tools don't want Canvas' navigation because
it is confusing to users to have 2 sets of prev/next
buttons. This allows Canvas' to be hidden in those
cases.

Test Plan:
 * the tool provider example on instructure's github has support for
   this in the extensions branch

 * send a post message of lti.scrollToTop from an LTI app
 * It should scroll to the top of the tool iframe

 * Send a post message lti.showModuleNavigation from an LTI app
 * The sequence footer should show/hide

Change-Id: I807119c3cadcbd04d90e68eeabdfe5872a3fffa4
Reviewed-on: https://gerrit.instructure.com/66439
Tested-by: Jenkins
Reviewed-by: Brad Horrocks <bhorrocks@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Nathan Mills <nathanm@instructure.com>