fix issue with jaws in ie for file previews
fixes CNVS-25082

There is this weird edge case in jaws where it won't properly hide
role=tree ul's. This makes sure only the file preview is being read by
the jaws screen reader in ie. Note, that jaws for some reason will let
you navigate to the <title> tag of the page and read it. As of now,
there is no fix for that.

Test Plan

Given I'm a user with Jaws and IE
And I'm previewing a file
When I navigate through the file preview
It should not read anything that is not in the preview overlay
Specifically, it should not read off the file tree

Change-Id: Id19b70ca7f3788c9690b15b2b306de0bf3a87445
Reviewed-on: https://gerrit.instructure.com/67332
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Ryan Allen <rallen@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>