submission preview_url should respect quiz_submission_version
fixes CNVS-25342

When constructing submission json for the API, the preview_url value was
always being supplied a version query param that pointed to the
version_number of the submission itself. Based on the behavior in the
submissions show action, it's clear that if the submission relates to a
quiz submission, what we actually want is to provide a version number
that corresponds to the submission's quiz_submission_version.

test plan:
- See repro steps on https://instructure.atlassian.net/browse/CNVS-25342
  for actual bug fix test instructions.

Change-Id: I3adb12ad349b93b351c0f9b14a03aad5150426ed
Reviewed-on: https://gerrit.instructure.com/68719
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
Tested-by: Jenkins
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: John Corrigan <jcorrigan@instructure.com>