fix tab issues on add module item dialog
fixes CNVS-25732

When you were on a select menu on this page and pressed enter, the tab
focus would disapear. This fixes two things on the page to solve the
problem.

1. It will disable the "Add Item" button by default, only enabling it
when an item has been selected.
2. Makes sure the focus doesn't change when pressing enter. If
something has been selected and enter is pressed, the form will now
submit like its suppose to.

Test Plan
Given you're on the modules page
And you have a module to add an item to
When you open the add item (+) dialog
You should see focus is on the "Assignments" select menu
When you press "enter"
Then focus should not change
When you select and item from the middle
And you press enter
Then the form should submit
And that item should be added to the module

Change-Id: Iec7427f68ae8c7037e607ef9fb30710a08f65f9a
Reviewed-on: https://gerrit.instructure.com/69154
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Jon Willesen <jonw@instructure.com>