disable auto focus when dialog pulls up
fixes CNVS-25967

This dialog has a tab with two options in it. It was auto focusing to
the first input in the dialog however it shouldn't be doing this. Now
it just focuses on the first tab item.

Test Plan
Given you're on the user settings page
And you have click the "Add contact method" link
When the dialog pulls up
You should see tabs with a form
And you should NOT be automatically focused on the first input field

Change-Id: If3a8f7b552a79e0ee94f7639f9785d7e90571c83
Reviewed-on: https://gerrit.instructure.com/69314
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>