Fixes over 30 character error for assignment names with symbols
Fixes: SIS-1641

Notes: Before if you had an assignment that had a name close to
30 characters and you had a symbol in the name, i.e. 'Buddy & Guy'
it would error on the assignment name validation for having a name
too long because the '&' would come in as '&amp;' causing the name
to go over 30 characters. In this commit you can have a name with
any symbol be valid if it's under 30 characters.

Change-Id: Ia136edaf9eeee290ef458662d27d3aab0842f72e
Reviewed-on: https://gerrit.instructure.com/71361
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
QA-Review: Kausty Saxena <kausty@instructure.com>
Product-Review: Ken Romney <kromney@instructure.com>