Add more accessible icon for post to SIS status
The previous icon for displaying whether an assignment in the
assignments list view is set to post to SIS wasn't accessible for
colorblind people in that it changed color but not shape. This has now
been replaced with a slightly different icon now so that the shape as
well as color change when it is toggled.

Fixes: SIS-1740

Test plan:
    Look at the list assignments view.

    Toggle the "post to sis" button for some assignments.

    The new icon is filled in with a clear "S"; the old icon is not
    filled in with a solid "S". See the Jira ticket for more info.

    Confirm all other accessible cases work as expected

Change-Id: Iff4f9bac77081959fb8c8cfb55181f7b3c9ce46a
Reviewed-on: https://gerrit.instructure.com/72597
Reviewed-by: Jayce Higgins <jhiggins@instructure.com>
Tested-by: Jenkins
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Ken Romney <kromney@instructure.com>