ensure overrides exist when has_overrides is true
when sorting assignments by due date in the
gradebook, we check to see if the assignments
have overrides. this is done by checking the
'has_overrides' property on an assignment.
sometimes an assignment's 'has_overrides' property
will return true even though it does not have any
overrides (we think this is caused by CNVS-26415),
and this causes the gradebook-spinner to run
forever. to fix this, we now make sure that an
assignment actually has overrides when its
'has_overrides' property is set to true.

closes CNVS-27409

test plan:
- get an assignment in a state such that its
  'has_overrides' property is true, but it
  doesn't actually have any overrides (i think
  you can do this by creating an assignment,
  assigning it to a single student, and then
  deleting that student)
- verify you can sort assignments by due date
  in the gradebook

Change-Id: I1f8f8f62e43da9c74feed3423031053bb4fb774c
Reviewed-on: https://gerrit.instructure.com/72718
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Tested-by: Jenkins
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>