hide assignment cog menu for non manage roles
fixes CNVS-9516

For account level roles that don't have the manage permission, you
should not be able to change the weight etc via the cog menu on the
assignments page. This commit ensures only users that have the manage
permission can see the cog.

Test Plan
Setting Up a User -----
1. Create a new account role (something like TestAdmin)
2. Set permissions on that role to manage and view assignments but NOT
manage a course
3. Create a new user from the admin section with that role
4. Become that user to test this
------

Given you are an account level user with the mangage_courses permission
set to true
When you go to the assignments page
Then you should see a cog menu with the weight form in it

Given you are an account level user with the mangage_courses permission
set to false
When you go to the assignments page
Then you should NOT see an admin cog menu with the weight form in it

Change-Id: I64179f247e83647b11f7f1d34358d9d31023f2e8
Reviewed-on: https://gerrit.instructure.com/72919
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Reviewed-by: Mike Nomitch <mnomitch@instructure.com>
Tested-by: Jenkins
Product-Review: Sterling Cobb <sterling@instructure.com>