handle bundler being installed to system gems
because we control what version of bundler is installed, we uninstall
the version provided by the base image and install the version we
expect to be there. however, if the base image installs bundler to a
system gem directory, an extra option must be provided to the
uninstall command. this is an issue as we are attempting to have the
base image install bundler system-wide.

Change-Id: I0b21e557315ed4ce7382efe3d740b602b016054d
Reviewed-on: https://gerrit.instructure.com/75328
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
Product-Review: Mark Severson <markse@instructure.com>
QA-Review: Mark Severson <markse@instructure.com>