Sessionless launch uses correct resource_link_id for assignment launches
Fixes PLAT-1396

We were not setting the link_code before initializing
the LtiOutboundAdapter, so it was defaulting to the
resource_link_id for the course.

The sessioned version of this launch already works
correctly. The implementation is in application_controller.rb:~1351,
and is tested in application_controller_spec.rb:~430.

Test Plan:
As an assignment tool to a course and create an assignment for it.
Use the api to generate a sessionless launch url.
That request looks like:
canvas.docker/api/v1/courses/1/external_tools/sessionless_launch?assignment_id=1&launch_type=assessment
In firefox (not chrome), open browser dev tools and
follow the link that was returned in the api request.
Notice in the POST to the lti tool that the
resource_link_id is no longer the same as the context_id.

Change-Id: I71cd944a38eb92dbd38f31d74b38e3117e5b67f9
Reviewed-on: https://gerrit.instructure.com/78240
Tested-by: Jenkins
Reviewed-by: Andrew Butterfield <abutterfield@instructure.com>
Reviewed-by: Nathan Mills <nathanm@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Karl Lloyd <karl@instructure.com>