Fix overlap on header toolbar and right sidebar in Discussions
closes: CNVS-26533

Test Plan:
- Go to discussions and add enough to be able to scroll the page and have
the toolbar stick to the top to follow you
- Toolbar should now overlap the right side RCE

Note: the right side no longer sticks with you down the page, but is
pinned to the top, so you will need to scroll up to view it. This was
part of another ticket.

Change-Id: I680b9ccc36502423eedcd757956503cac8adf724
Reviewed-on: https://gerrit.instructure.com/78737
Tested-by: Jenkins
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>