spec: publish first provisional grade if none selected
when a moderator does not "Select" a reviewer and
then "Post"s grades for an assignment, the grade
and rubric comments are lost and are not visible
in Speedgrader. now, when a moderator does not
"Select" a reviewer and then "Post"s grades
for an assignment, the first reviewer's grade and
rubric comments will be used and will be visible
in Speedgrader. this bug is fixed by bf1794df. this
commit adds specs only.

closes CNVS-29116

test plan:
 - create an assignment and allow moderated grading
 - click the '+rubric' button and link the assignment
   to a rubric, allowing free-form comments. also check
   the box to use the rubric to grade the assignment
 - click the 'moderate' button on the assignment page
 - click the checkbox to select all students and then
   click the '+ reviewer' button
 - go to Speedgrader
 - for a student, view the rubric and leave some free-form
   comments and give them a grade and click save
 - go back to the moderation page (make sure you refresh the
   page if you haven't since you went into Speedgrader).
   without "Select"ing any reviewers, "Post" the assignment
 - go to Speedgrader and view the submission for the student.
   you should see the grade and rubric comments that the reviewer
   left

Change-Id: I200d6d951338cdfc01e8763c720cd4ed5dc7cd12
Reviewed-on: https://gerrit.instructure.com/79554
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Neil Gupta <ngupta@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>