spec: make spec that tests that turning off inbox works in newUI
so this is actually a legitimate failure in New UI.
if you turn off the inbox, it should not show that
link in the header, but we still do in newUI.
if we fix the code this test will pass.

test plan:
we can either add code to properly hide the inbox
as a new patchset to this commit, if that's the case:

* run CANVAS_FORCE_USE_NEW_STYLES=1 bin/rspec spec/selenium/conversations/conversations_inbox_opt_out_spec.rb
* it should pass

or we can just commit this as-is and have it there
reminding us that we need to fix that in new ui.
in which case:
* run CANVAS_FORCE_USE_NEW_STYLES=1 bin/rspec spec/selenium/conversations/conversations_inbox_opt_out_spec.rb
* it should fail
* but the mandatory jenkins builds should pass
  since it still works in oldUI, so we can commit it

Change-Id: Idcd600551fb4e093b1d9eda41b55955873a28a2c
Reviewed-on: https://gerrit.instructure.com/80532
Tested-by: Jenkins
Reviewed-by: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Matt Fairbourn <mfairbourn@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>