theme editor for siteadmin
fixes CNVS-28289

this enables theme editor to work as a virtual "parent" account for root
account themes, but only root accounts on the same shard as the site
admin account.  this is a limitation we hope to remove in the future,
but replicating the site admin theme to all shards is going to be
significantly more complicated.

test plan:
- create a mixture of root accounts on the same shard as site admin and
different shards
- regression test theme editor in siteadmin.  it should act as a parent
for same-shard accounts and not interact with different-shard accounts

Change-Id: Ia7e3e1a8adbb7bc2e49b2d9b5ef3b4d30cf4c581
Reviewed-on: https://gerrit.instructure.com/82000
Reviewed-by: Cody Cutrer <cody@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins
QA-Review: Benjamin Christian Nelson <bcnelson@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>