focus on the close button when dialog opens
fixes CNVS-29848

When you click the edit icon when moderating students in a quiz, the
initial focus on the dialog should be on the close link in the top
right corner. This commit sets that focus after the dialog loads.

Test Plan

Given you are a teacher
And you have a quiz
And a student is in the course with the quiz
When you go to the moderate quiz page
And you click on the pencil icon next to a student to moderate them
Then a dialog should appear
And inital focus should be set on the top right close link
And this should work for screen readers as well

Change-Id: Iebd348f5022be4004d82dc24b99ea4964e955988
Reviewed-on: https://gerrit.instructure.com/82670
Reviewed-by: Mike Nomitch <mnomitch@instructure.com>
Tested-by: Jenkins
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>