default to first provisional grade when posting grades
When posting grades to a moderated assignment, if a second
reviewer is assigned to a student but didn't assess the student,
the moderated grading page shows the first reviewer's score
selected and being posted to the assignment, but in actuality no
score, comments, or annotations are being saved for that student.

This fixes that.

fixes CNVS-28755

test plan:
 1. Create a moderated grading assignment
 2. Use SpeedGrader™ to enter a score for a student
 3. Visit the 'Moderate Assignment' page
 4. Add a second reviewer for that student
 5. Click "Post" to post the grades
 6. Notice the "Grade" column displays the grade of the first reviewer
 7. Visit SpeedGrader™ for that student
 8. Verify that the score matches the one provisional score entered

Change-Id: I687ea43ac00abdcf21dedbcab62304ff1e58e622
Reviewed-on: https://gerrit.instructure.com/82677
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Jenkins
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>