spec: remove redundant check, make it less flaky
this spec was checking the same page twice (via two different links),
and a different student could sometimes be selected the second time,
causing the obfuscated names to be switched.

when student names are hidden, students are sorted by submissions.id
alphabetically. so if the ids are 9 and 10 (or 99 and 100), the second
one will come first in the dropdown (this happens 2.3% of the time,
depending on which specs run prior). on the first page load, "first
student" was always selected since the url was set prior to hiding
student names. on the second page load, "second student" could be
selected if the ids were as explained above, causing the spec to fail

Change-Id: I1e75ad490acd87912ec4eed90f456c1992211001
Reviewed-on: https://gerrit.instructure.com/89482
Tested-by: Jenkins
Reviewed-by: Shawn Meredith <shawn@instructure.com>
Product-Review: Shawn Meredith <shawn@instructure.com>
QA-Review: Shawn Meredith <shawn@instructure.com>