ruby 2.3: fix marshalling problem
in ruby 2.1, AllowedServicesHash.reject {} will return a AllowedServicesHash,
on 2.3 it will return a plain Hash. so dup it instead (which returns an
AllowedServicesHash), and then reject! (which doesn't change the type
of the object). this works in both 2.1 and 2.3

Change-Id: Ic6eac219e31cd469afd45d026822654c11180b49
Reviewed-on: https://gerrit.instructure.com/92553
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>