spec: fix failing movetodialog spec
Because of the async nature of the callbacks in the modal
this was failing because the spec ran too quickly.  This
wraps it in a simple 1ms setTimeout that causes it to wait
a tick, then make the remaining asserts asyncronously.

Test Plan:
  - Specs pass

Change-Id: Ied07f1ef71cf183f49bcb702ca886edeef1ad4de
Reviewed-on: https://gerrit.instructure.com/97460
Reviewed-by: Michael Brewer-Davis <mbd@instructure.com>
Tested-by: Jenkins
Product-Review: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clay Diffrient <cdiffrient@instructure.com>