Add DUE_DATE_REQUIRED_FOR_ACCOUNT to the js_env
Added DUE_DATE_REQUIRED_FOR_ACCOUNT to the js_env in the assignments
controller.  This will be used in to verify if a date is required for
the assignment, and this functionality will appear in the next commit.

Refs: SIS-2636

Test Plan: QA will review the functionality in the final commit.

Change-Id: I320337aa6fc8779adb5e022b6d846736e603da69
Reviewed-on: https://gerrit.instructure.com/101392
Tested-by: Jenkins
Reviewed-by: Ken Romney <kromney@instructure.com>
Product-Review: Stewie aka Nicholas Stewart <nstewart@instructure.com>
QA-Review: Stewie aka Nicholas Stewart <nstewart@instructure.com>