transform locked item links; lock item export fix
for web zip exports, do the link transform even if the item is locked.

bug fix: if an item is locked for the user doing the CC export, that
item should not be exported.

refs OFFW-136

test plan:
* items locked by modules or by date should not have their content
exported in the web zip export, even if the item is not in a module.
* links to locked items in item content should be transformed into local
links the same way as links to unlocked content are transformed.

Change-Id: I7017f6fa489444db6bf519ad7fcbe527057fe471
Reviewed-on: https://gerrit.instructure.com/105786
Tested-by: Jenkins
Reviewed-by: James Williams  <jamesw@instructure.com>
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>