timeout when editing grading periods sets with many courses
The issue occured on a term with a large number of courses (~2000)
when a grading period set is attached or detached from a term,
or is attached to a term and its weighted flag changes.  These
events trigger grade recalculation on all assignments in the term.
While the grade recalculation jobs each run in the background,
the loop which submits those jobs runs in the foreground.

To fix this issue, the loop that enqueues one job for each course
should itself be a background job, if running in production mode.

fixes CNVS-36070
[ci-coverage]

Test Plan
  1. Seed local canvas with 100 courses, each with a few submissions
     Go to admin Grading window, edit grading period
     Use browser network timing window to measure response time

  100 Courses, simulating production (delayed job)
	change weighted flag: 0.21s, 0.30s
	detach term from gpg: 0.13s, 0.13s
	attach term to gpg: 0.15s, 0.17s

  100 Courses, not simulating production (immediate)
	change weighted flag: 4.11s
	detach term from gpg: 7.95s
	attach term to gpg: 5.06s

     In each case, verify that scores were recalulated properly

Change-Id: I08674d3871d0d70c1f5ea569e241343c7d9ca6d8
Reviewed-on: https://gerrit.instructure.com/111657
Tested-by: Jenkins
Reviewed-by: Derek Bender <djbender@instructure.com>
Reviewed-by: Neil Gupta <ngupta@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>